+++ src/ScatterChartThree.js
@@ -403,13 +403,13 @@ function ScatterChartThree(props) {
                     lastHoverIndexRef.current = selectedIndex;
                     if (selectedIndex !== -1) {
                         let value = trace.values[selectedIndex];
-                        let categoryObject = props.categoricalNames[trace.name];
-                        if (categoryObject) {
-                            let renamedValue = categoryObject[value];
-                            if (renamedValue != null) {
-                                value = renamedValue;
-                            }
+                        let categoryObject = props.categoricalNames[trace.name] || {};
+
+                        let renamedValue = categoryObject[value];
+                        if (renamedValue != null && renamedValue.newValue != null) {
+                            value = renamedValue.newValue;
                         }
+
 
                         if (typeof value === 'number') {
                             value = numberFormat2f(value);