+++ docs/release_notes.rst
@@ -1,6 +1,10 @@
 Release Notes
 -------------
 
+- 1.1.20.post1 `September 8, 2021`
+    * Save categorical colors to database
+    * Save cluster positive and negative markers to database
+
 - 1.1.20 `September 2, 2021`
     * Added static website hosting capabilities
     * Performance improvements